composer/backend/api/app/requirements.txt
fastapi
uvicorn
python-dotenv
pymysql